Move mistakenly placed tool lease
